feat: 重构聊天历史压缩与提醒调度系统，新增 BT 期货交易插件
🎯 核心改进:
- AI 聊天历史压缩：实现对话历史自动摘要压缩，优化上下文长度管理
- 提醒调度优化：从轮询模式改为 Next-Fire 按需唤醒模式，降低资源消耗
- UI 界面重构：优化主窗口布局，改进标题栏和设置页面交互体验

📦 新增功能:
- TokenTrackingChatClient: Token 使用量追踪客户端
- BT 期货交易插件：完整的期货交易解决方案（含文档和测试）
- WebSocket 桥接插件方案文档
- VS Code 工作空间配置

🔧 技术重构:
- ChatHistoryDataProvider: 移除冗余依赖，引入 IServiceProvider 实现延迟解析
- ReminderScheduler: 实现动态重调度机制，支持列表变更即时响应
- MainWindow: 重构 AXAML 布局，优化控件层次结构和样式
- 实体类扩展：AiModelEntity/AiProviderEntity/ChatSessionEntity 新增字段

📝 文档更新:
- 期货交易插件解决方案文档
- 期货交易插件 v1 合约文档
- WebSocket 桥接插件方案文档

🐛 修复问题:
- 修复聊天历史记录保存逻辑
- 修复提醒调度器资源释放问题
- 修复 UI 主题切换和布局渲染问题

📊 变更统计:
- 19 个文件修改，776 行新增，625 行删除
- 6 个新文件/目录添加

+++ Src/Netor.Cortana.Entitys/Entities/ChatSessionEntity.cs
@@ -65,6 +65,6 @@ namespace Netor.Cortana.Entitys
         /// <summary>
         /// 该会话累计消耗的总令牌数，用于统计和展示 API 用量。
         /// </summary>
-        public int TotalTokenCount { get; set; }
+        public long TotalTokenCount { get; set; }
     }
 }
